update the provision script to install vim when it has not been configured yet. Install node, npm and grunt manually instead of using package manager.

+++ vagrant/bootstrap.sh
@@ -9,20 +9,22 @@ apt-get install -y subversion
 apt-get install -y fontconfig
 
 # Set up Vim and Tmux
-export HOME="/home/vagrant"
-cd $HOME
-rm -Rf $HOME/dotfiles
-rm -Rf $HOME/molokai
-rm -Rf $HOME/.vim/colors
-git clone https://github.com/rudylee/dotfiles $HOME/dotfiles
-git clone https://github.com/gmarik/vundle.git $HOME/.vim/bundle/vundle
-git clone https://github.com/tomasr/molokai.git $HOME/molokai
-ln -sf $HOME/dotfiles/vimrc $HOME/.vimrc
-mkdir $HOME/.vim/ftplugin
-ln -sf $HOME/dotfiles/ftplugin/php.vim $HOME/.vim/ftplugin/php.vim
-ln -sf $HOME/dotfiles/tmux.conf $HOME/.tmux.conf
-mv $HOME/molokai/colors $HOME/.vim
-vim +BundleInstall +qall
+if ! type "tmux" > /dev/null; then
+    export HOME="/home/vagrant"
+    cd $HOME
+    rm -Rf $HOME/dotfiles
+    rm -Rf $HOME/molokai
+    rm -Rf $HOME/.vim/colors
+    git clone https://github.com/rudylee/dotfiles $HOME/dotfiles
+    git clone https://github.com/gmarik/vundle.git $HOME/.vim/bundle/vundle
+    git clone https://github.com/tomasr/molokai.git $HOME/molokai
+    ln -sf $HOME/dotfiles/vimrc $HOME/.vimrc
+    mkdir $HOME/.vim/ftplugin
+    ln -sf $HOME/dotfiles/ftplugin/php.vim $HOME/.vim/ftplugin/php.vim
+    ln -sf $HOME/dotfiles/tmux.conf $HOME/.tmux.conf
+    mv $HOME/molokai/colors $HOME/.vim
+    vim +BundleInstall +qall
+fi
 
 # Set up the bashrc
 if ! grep -q "export TERM=xterm-256color" "$HOME/.bashrc"; then
